Migrate Jaeger integ tests to integ tests
